Add isatty and the ability to set a custom scheme.
Now pgg checks whether the stdout is a tty and if so only prints the
body of the request allowing it to be piped to files or pagers.
Plus now you can set a custom protocol scheme in the environments and if
it's not found in the url it's added with the default out of the box.